Updated to the newest version of pdsdata and tried to write wavelength information to HDF5 files also

+++ LCLS/pdsdata/xtc/XtcFileIterator.hh
@@ -2,7 +2,7 @@
 #define Pds_XtcFileIterator_hh
 
 #include "pdsdata/xtc/Dgram.hh"
-#include <new>
+
 #include <stdio.h>
 
 namespace Pds {
@@ -14,7 +14,7 @@ public:
   Dgram* next();
 private:
   FILE*    _file;
-  unsigned _maxDgramSize;
+  size_t   _maxDgramSize;
   char*    _buf;
 };
